feature: prompt user when delete

+++ SmartHomeManager/SmartHomeManager.View/src/components/Rules/Buttons.js
@@ -2,8 +2,12 @@ import { Button, Td } from "@chakra-ui/react";
 import React, { useState } from "react";
 import { Link } from "react-router-dom";
 import axios from "axios";
+
+import ConfirmationScreen from "../../pages/rules/ConfirmationScreen";
+
 function Buttons({props, deleteRule,editButton}) {
 	const [isDeleted, setIsDeleted] = useState(false);
+	const [showConfirmation, setShowConfirmation] = useState(false);
 	console.log(props);
 	return (
 		<Td>
@@ -14,13 +18,24 @@ function Buttons({props, deleteRule,editButton}) {
 				ml={2}
 				colorScheme="red"
 				onClick={() => {
-					deleteRule(props.ruleId).catch((error) => {
-						console.error(error);
-					});
+					setShowConfirmation(true);
 				}}
 			>
 				Delete
 			</Button>
+			{showConfirmation && (
+				<ConfirmationScreen
+					onConfirm={() => {
+						deleteRule(props.ruleId).catch((error) => {
+							console.error(error);
+							});
+						setShowConfirmation(false);
+					}}
+					Close={()=>{
+						setShowConfirmation(false);
+					}}
+				/>
+			)}
 		</Td>
 	);
 }
